+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - function mispelling in rgb widget
 - bug fixes
   - zeroconf: unpublish when shutting down
+  - ios 9.3 compatibility issue
 - widgets
   - modal: added `popupLeft` and `popupTop` properties
   - faders/meter: add css variable `--gauge-opacity` to allow setting a fixed gauge opacity for meters and compact faders